Share props with home page and home seo
