+++ packages/gatsby-theme-store/src/components/SearchPage/SEO.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import type { FC } from 'react'
 
-import SiteMetadataSEO from '../HomePage/SEO'
+import SiteMetadata from '../SEO/SiteMetadata'
 import type { SearchPageProps } from '../../templates/search'
 
 const SEO: FC<SearchPageProps> = ({ data }) => (
-  <SiteMetadataSEO title={data.vtex.productSearch!.titleTag!} />
+  <SiteMetadata title={data.vtex.productSearch!.titleTag!} />
 )
 
 export default SEO
